# openapi.yaml
...
openapi: 3.0.3
info:
    title: CompleteExampleService API
    description: 完整功能展示服务 - HTTP方法、参数类型、验证规则的综合示例
    version: 0.0.1
paths:
...
    /api/v1/users:
        get:
            tags:
                - CompleteExampleService
            description: 简单GET请求 - 查询参数
            operationId: CompleteExampleService_ListUsers
            parameters:
                - name: page
                  in: query
                  description: 分页参数 - Form绑定 + 验证
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: sortBy
                  in: query
                  description: 排序参数 - 枚举验证
                  schema:
                    type: string
                - name: sortOrder
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  description: 过滤参数 - 数组
                  schema:
                    type: array
                    items:
                        type: string
                - name: roles
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
                - name: includeDeleted
                  in: query
                  description: 布尔参数
                  schema:
                    type: boolean
                - name: includeStats
                  in: query
                  schema:
                    type: boolean
                - name: createdAfter
                  in: query
                  description: 日期范围 - 日期格式验证
                  schema:
                    type: string
                - name: createdBefore
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListUsersResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - CompleteExampleService
            description: POST请求 - JSON Body
            operationId: CompleteExampleService_CreateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateUserResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - CompleteExampleService
            description: DELETE请求 - 批量操作
            operationId: CompleteExampleService_BatchDeleteUsers
            parameters:
                - name: userIds
                  in: query
                  description: 查询参数 - 批量删除
                  schema:
                    type: array
                    items:
                        type: string
                - name: hardDelete
                  in: query
                  schema:
                    type: boolean
                - name: deleteReason
                  in: query
                  schema:
                    type: string
                - name: batchConfirmation
                  in: query
                  description: Header参数 - 批量操作安全确认
                  schema:
                    type: string
                - name: authorization
                  in: query
                  schema:
                    type: string
                - name: operationId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchDeleteUsersResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /api/v1/users/search:
        get:
            tags:
                - CompleteExampleService
            description: GET请求 - 复杂查询 + Header参数
            operationId: CompleteExampleService_SearchUsers
            parameters:
                - name: query
                  in: query
                  description: 查询参数 - 必需 + 长度验证
                  schema:
                    type: string
                - name: searchFields
                  in: query
                  description: 搜索选项
                  schema:
                    type: array
                    items:
                        type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: clientId
                  in: query
                  description: Header参数 - 客户端信息
                  schema:
                    type: string
                - name: requestId
                  in: query
                  schema:
                    type: string
                - name: userAgent
                  in: query
                  schema:
                    type: string
                - name: apiKey
                  in: query
                  schema:
                    type: string
                - name: latitude
                  in: query
                  description: 地理位置过滤 - 数值范围验证
                  schema:
                    type: number
                    format: double
                - name: longitude
                  in: query
                  schema:
                    type: number
                    format: double
                - name: radiusKm
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: minAge
                  in: query
                  description: 高级过滤
                  schema:
                    type: integer
                    format: int32
                - name: maxAge
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: country
                  in: query
                  schema:
                    type: string
                - name: city
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SearchUsersResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/users/{userId}:
        get:
            tags:
                - CompleteExampleService
            description: GET请求 - 路径参数
            operationId: CompleteExampleService_GetUser
            parameters:
                - name: userId
                  in: path
                  description: 路径参数 - UUID验证
                  required: true
                  schema:
                    type: string
                - name: fields
                  in: query
                  description: 查询参数 - 选择字段
                  schema:
                    type: array
                    items:
                        type: string
                - name: includeProfile
                  in: query
                  schema:
                    type: boolean
                - name: includePosts
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetUserResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - CompleteExampleService
            description: PUT请求 - 完整更新
            operationId: CompleteExampleService_UpdateUser
            parameters:
                - name: userId
                  in: path
                  description: 路径参数
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateUserResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - CompleteExampleService
            description: DELETE请求
            operationId: CompleteExampleService_DeleteUser
            parameters:
                - name: userId
                  in: path
                  description: 路径参数
                  required: true
                  schema:
                    type: string
                - name: hardDelete
                  in: query
                  description: 查询参数 - 删除选项
                  schema:
                    type: boolean
                - name: deleteReason
                  in: query
                  schema:
                    type: string
                - name: transferData
                  in: query
                  schema:
                    type: boolean
                - name: transferToUser
                  in: query
                  schema:
                    type: string
                - name: confirmation
                  in: query
                  description: Header参数 - 安全确认
                  schema:
                    type: string
                - name: authorization
                  in: query
                  schema:
                    type: string
                - name: adminToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteUserResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        patch:
            tags:
                - CompleteExampleService
            description: PATCH请求 - 部分更新
            operationId: CompleteExampleService_PatchUser
            parameters:
                - name: userId
                  in: path
                  description: 路径参数
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PatchUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PatchUserResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/users/{userId}/posts:
        post:
            tags:
                - CompleteExampleService
            description: POST请求 - 混合参数 (路径 + 查询 + Body + Headers)
            operationId: CompleteExampleService_CreatePost
            parameters:
                - name: userId
                  in: path
                  description: 路径参数
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreatePostResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        CreateUserRequest:
            type: object
            properties:
                username:
                    type: string
                    description: 基本信息 - 必需字段 + 格式验证
                email:
                    type: string
                password:
                    type: string
                fullName:
                    type: string
                    description: 个人信息 - 可选字段
                phone:
                    type: string
                age:
                    type: integer
                    format: int32
                gender:
                    type: string
                bio:
                    type: string
                address:
                    allOf:
                        - $ref: '#/components/schemas/Address'
                    description: 地址信息 - 嵌套对象验证
                hobbies:
                    type: array
                    items:
                        type: string
                    description: 数组字段 - 数组长度验证
                languages:
                    type: array
                    items:
                        type: string
                socialLinks:
                    type: object
                    additionalProperties:
                        type: string
                    description: Map字段
                preferences:
                    type: object
                    additionalProperties:
                        type: string
                settings:
                    allOf:
                        - $ref: '#/components/schemas/UserSettings'
                    description: 用户设置
                agreeTerms:
                    type: boolean
                    description: 验证字段
                subscribeNewsletter:
                    type: boolean
                referralCode:
                    type: string
                    description: 自定义标签示例
                tags:
                    type: array
                    items:
                        type: string
...
        ListUsersResponse:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/User'
                totalCount:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                hasNext:
                    type: boolean
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...